str(Exception) doesn't work for anything but ASCII Exception texts, so '%s' % exception_instance is unsafe unless exception_instance.args contains only byte strings.
The change in alltests.py is a first attempt to catch such cases where
a str(Exception) is done with an Exception text which is not
necessarily ASCII-only (i.e. with a unicode string in Exception.args),
but since most input data (in the totest dicts) is passed in as byte
strings, it doesn't catch much.


git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@3253 929543f6-e4f2-0310-98a6-ba3bd3dd1d04