Added the rest of the new LaTeX writer (I accidentally committed newlatex.py together with nodes.py).
To test the new LaTeX writer you must (currently) specify
tools/latex.tex as a stylesheet.  You can use
test/functional/input/standalone_rst_latex.txt as input document.

Note that the writer is not yet complete; I'm checking it in to the
trunk anyway because I don't want to have to merge changes from the
trunk to a branch.  Things to do:

* Implement support for missing elements.
* Restructure.
* Add tests for a lot of special cases.
* Write documentation, esp. how to customize the behavior.


git-svn-id: https://docutils.svn.sourceforge.net/svnroot/docutils/trunk@3260 929543f6-e4f2-0310-98a6-ba3bd3dd1d04