Added syntax highlighting for literal code blocks.
2006/12/19 -- Version 1.0b
--------------------------

Added syntax highlighting for literal code blocks.  Syntax
highlighting is applied only if Pygments is installed and the
--add-syntax-highlighting command line flag is used.  Pygments can
be found here:  http://pygments.pocoo.org/.  See the odtwriter
documentation for information about the styles used for syntax
highlighting.


2006/12/18 -- Version 1.0b
--------------------------

Fixed zipfile so that members of .odt have UNIX access permissions
and are stored deflated.



git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@4835 929543f6-e4f2-0310-98a6-ba3bd3dd1d04