Clear up and document patch [ 120 ].
The colwidths argument for <tgroup> violates the docutils.dtd.
As the "auto"/"given" tags may be ignored by the writers, they may be
stored as classes instead.

The default case (grid for tables, equal widths for ListTable and CSVTable
does not need/get a tag. This way, writers can decide whether to leave
column width determination to the backend when the class "colwidths-auto" is
set (opt-in) or by default (opt-out with "colwidths-given").

git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@7958 929543f6-e4f2-0310-98a6-ba3bd3dd1d04