Don't drop "code" directive content if Warnings are silenced.
Output the content without syntax highlight to prevent
data loss.

It is still recommended to solve the issue instead of just
silencing the warning. (Languages that are unknown to the
Pygments Lexer should be specified as :class: option value
instead of directive argument.)

Thanks to Jérôme Carretero for the report and suggestion.

git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@8596 929543f6-e4f2-0310-98a6-ba3bd3dd1d04