HTML5: Improve accessibility of section headings.
Mark the internal back-refences from section headings to the ToC
with DPub-ARIA role "doc-backref".

Use "aria-level" instead of a class argument value for section headings
with nesting level > 6 (for level 1 to 6, the level is determined by
the HTML tag).
Mind that in HTML5, heading level 1 is reserved for the document title
and section headings start with level 2.

Add functional test sample for deeply nested sections.
Merge functional tests for additonal HTML5 features.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@8786 929543f6-e4f2-0310-98a6-ba3bd3dd1d04