Node.traverse() returns a list again to restore backwards compatibility.
Although the documented API only promised an "iterable" as return
value, the implementation returned a list.
Because Sphinx < 4.0.0 relies on this and Sphinx < 3.5.4 places no upper
limit to the Docutils dependency, this led to many errors "in the wild".

Fixes bug #431.

New method Node.findall(): like Node.traverse() but returns an iterator.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@8885 929543f6-e4f2-0310-98a6-ba3bd3dd1d04