Avoid mutables as function argument defaults.
Mutable default values for functions are considered an anti-pattern
(cf. bug #430 and
https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments).

Mutable default values (empty lists and dictionaries) were used
in the "interpreted role function" API:
Interpreted role functions must accept (optional)
"options" and "content" arguments and must not modify passed values.

OTOH, the actual default value is irrelevant for the API, so there is
no need to use mutable defaults in the specification.
As a precaution and example for robust coding, mutable defaults
are replaced with None in the function definitions.

The documentation is updated to the current implementation (including
changes in [r8254] to apply feature request #63).

Fixes bug #430.

git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@8919 929543f6-e4f2-0310-98a6-ba3bd3dd1d04