core.Publisher.publish(): Prompt when waiting for input from a terminal.
When front-end tools are used without arguments,
Docutils reads from stdin.

Currently, when a user calls a front-end without arguments,
there is no response until the user presses Ctrl-D (Ctrl-Z on Windows)
or aborts with Ctrl-C.

This commit changes the behaviour to tell the user what the front-end
expects and where to get more help.

git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@9043 929543f6-e4f2-0310-98a6-ba3bd3dd1d04