Import ``DocutilsTestSupport`` from ``test``
This is the first step of moving to the standard ``unittest`` style.
It involves making the ``test/`` directory a psuedo-package and
importing the ``DocutilsTestSupport`` module from it, instead of the
path manipulation previously carried out in ``__init__.py`` files.
As of this, we also blank out all ``__init__.py`` files in the test
directories.


git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@9140 929543f6-e4f2-0310-98a6-ba3bd3dd1d04