Fix failures when running "transform" test modules as "__main__".
Also restructure `test_docinfo` to reduce code duplication.

Running "transform" test modules independently failed due to an additional
INFO message:

Loading a language module results in an INFO message. Import statements
don't trigger the INFO message if the language module is already loaded.

When running a test module as __main__, the INFO message is added to the
first document using both, the language module and the `Messages` or
`TestMessages` transform. When running the test suite, the language
module may already be loaded by another test.
For stable results, the optional message must be filtered.

In "transform" tests, we do not load the default transforms, so
INFO messages are only filtered if `FilterMessages` is explicitly added.
As `FilterMessages` only acts on attached messages, it must run
*after* `TestMessages`, therefore we load it with custom priority 890.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@9321 929543f6-e4f2-0310-98a6-ba3bd3dd1d04