Refactor HTML writer functions for definition lists.
In HTML, the optional "term classifier" is nested in the `<dt>` element.

Up to now, the closing `<dt>` was added in `visit_definition()`.
This is, however, problematic if we want to allow a "definition_list_item" with
multiple "term"s (cf. [feature-requests:60]).
Now, `</dt>` is appended in `depart_term()`, but only if no "classifier" node
follows. `depart_classifier()` also calls `depart_term()` so that `</dt>` is
added after the last "classifier".

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@9549 929543f6-e4f2-0310-98a6-ba3bd3dd1d04