Doctree Documentation: Remove trailing semi-colon from parameter entity names.
When using parameter entity names in running text, the trailing semi-colon
clashes with the usual interpretation of this interpunctation.

While the semicolon is required when referencing a parameter entity,
it is not used in the entity definition.

The leading percent sign is kept as a marker for a parameter entity.

This nomenclature follows the example of the
"XML Exchange Table Model Document Type Definition"
https://www.oasis-open.org/specs/tm9901.html

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@9661 929543f6-e4f2-0310-98a6-ba3bd3dd1d04