Doctree validation: declare valid element children (part 2).
Declare valid children and valid number of children for all
document tree elements.

Requires some re-ordering (define sub-elements first so that
the class can be used in the content declaration of the parent).

New element category class `nodes.PureTextElement` for nodes that do
not accept Inline elements, only text (was a TODO comment).

Add declarations for valid attributes of table elements defined
in the Exchange Table Model but not used/supported by Docutils.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@9694 929543f6-e4f2-0310-98a6-ba3bd3dd1d04