xml-parser: Use "XMLPullParser" to add source line number to elements.
Use an "XMLPullParser" to register source line numbers in the parsed elements.
Transfer them to the internal "line" attribute of the document-tree nodes.

(An attempt to create doctree node based on "start"/"end" events became
too complicated, because tailing text is not yet in `element.tail`
when an element's "end" event is reported.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@9736 929543f6-e4f2-0310-98a6-ba3bd3dd1d04