Add type hints to ``docutils.nodes``
A significant number of parameters and attributes are typed as
expecting or yielding ``Element | Text`` rather than ``Node``,
which is the more obvious choice. This union type is used because
``Node`` is de facto an abstract base class --- it should not
be used or instantiated. Notably, ``Node.children`` is not defined,
``Node.attributes`` does not exist, etc. The two direct subclasses
fill in many of these missing pieces, making static typing both
more precise and easier.


git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@9813 929543f6-e4f2-0310-98a6-ba3bd3dd1d04