Make targets from hyperlink references with embedded URI or alias implicit.
Embedding a URI or alias name in a hyperlink reference is no explicit request
for a target (cf. bug #502). Downgrade the priority of the generated targets from
explicit to implicit to avoid surprising warnings in case of name duplication.

(Anonymous references can be used to avoid the warnings but, in case of a typo,
the warnings/errors have no source/line information.)

git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@10176 929543f6-e4f2-0310-98a6-ba3bd3dd1d04