rST parser: Use `section_level_offset` instead of `memo.section_level`.
Keeping record of the current section level in the state machine's "memo"
is cumbersome and error prone because it needs to be updated with every switch
of the current node.

Store the difference between the intended start level of nested parsing and the
number of parents of the base node in the new attribute `section_level_offset`.
Use it to correct the section level determined via `node.section_hierarchy()`.

git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@10229 929543f6-e4f2-0310-98a6-ba3bd3dd1d04